requirements.txt
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudinary==1.30.0
colorama==0.4.5
cryptography==38.0.1
Deprecated==1.2.13
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
gcloud==0.18.3
googleapis-common-protos==1.56.4
greenlet==1.1.3
gunicorn==20.1.0
httplib2==0.20.4
idna==3.4
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.4.2
MarkupSafe==2.1.1
oauth2client==4.1.3
protobuf==4.21.6
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
pyparsing==3.0.9
Pyrebase4==4.5.0
python-decouple==3.6
python-dotenv==0.21.0
python-jwt==3.3.5
requests==2.28.1
requests-toolbelt==0.9.1
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.41
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
